Update : entity Default 값 세팅 에러 해결

+++ src/main/java/com/alzal/nadeulseoulbackend/domain/mypage/dto/MypageInfoDto.java
@@ -9,6 +9,6 @@ public class MypageInfoDto {
     private Long userSeq;
     private String nickName;
     private String emoji;
-    private Long followeeCount;
-    private Long followerCount;
+    private Integer followeeCount;
+    private Integer followerCount;
 }
